requirements.txt
opencv-python
gradio
gradio-webrtc
mediapipe
Pillow